requirements.txt
streamlit==1.32.0
python-dotenv==1.0.0
requests==2.31.0
uuid==1.30
monaco-editor==0.0.1
streamlit-monaco==0.1.3 